Refactor resources into package with panel, ocr, and cache modules
